+++ src/components/ShareGame.jsx
@@ -1,0 +1,26 @@
+import React, { useState } from 'react';
+import { historyToActions, actionsToQueryString } from '../url-query-state';
+
+export default function ShareGame({ gameState }) {
+  const [urlToShare, setUrlToShare] = useState('');
+  function displayUrlToShare() {
+    const query = actionsToQueryString(historyToActions(gameState.history));
+    const { protocol, host, pathname } = window.location;
+    const urlWithoutQuery = `${protocol}//${host}${pathname}`;
+    setUrlToShare(`${urlWithoutQuery}?${query}`);
+  }
+
+  return (
+    <>
+      <button type="button" className="button" onClick={displayUrlToShare}>
+        Share game
+      </button>
+      <input
+        type="text"
+        readOnly
+        data-testid="urlShareInput"
+        value={urlToShare}
+      />
+    </>
+  );
+}
