+++ src/components/UltimateGame.jsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 import Board from './Board';
 import Status from './Status';
 import SaveAndLoad from './SaveAndLoad';
+import ShareGame from './ShareGame';
 import TimeTravelButton from './TimeTravelButton';
 import { getColorClass } from '../helpers';
 import { ultimateTicTacToe, calculateUltimateWinner } from '../ultimate-game';
 import { playSquare, timeTravel, toggleSpecialIcons } from '../actions';
-import { historyToActions, actionsToQueryString } from '../url-query-state';
 
 export default function UltimateGame() {
   const [state, setState] = useState(ultimateTicTacToe(undefined, ''));
@@ -83,30 +83,6 @@ export default function UltimateGame() {
           )}
         </ol>
       </div>
-    </>
-  );
-}
-
-function ShareGame({ gameState }) {
-  const [urlToShare, setUrlToShare] = useState('');
-  function displayUrlToShare() {
-    const query = actionsToQueryString(historyToActions(gameState.history));
-    const { protocol, host, pathname } = window.location;
-    const urlWithoutQuery = `${protocol}//${host}${pathname}`;
-    setUrlToShare(`${urlWithoutQuery}?${query}`);
-  }
-
-  return (
-    <>
-      <button type="button" className="button" onClick={displayUrlToShare}>
-        Share game
-      </button>
-      <input
-        type="text"
-        readOnly
-        data-testid="urlShareInput"
-        value={urlToShare}
-      />
     </>
   );
 }
